requirements.txt
librosa==0.9.1
loss-landscapes==3.0.6
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
plotly==5.18.0
PyYAML==6.0.1
scipy==1.7.3
tensorboardX==2.5
torch==1.13.1
tqdm==4.67.1